(maint) Fixed formatting of several files
Corrected the usage of the following blocks:

* if

There are other blocks that also need to be updated, by these are
within comments on the generated code, which flows through from Cake
itself.  These have been addressed in this PR:

https://github.com/cake-build/cake/pull/4519

And will cause unit test failures when those are brought through.